Move headers to mega65 directory
