+++ include/mega65/fileio.h
