+++ include/mega65/hal.h
