+++ include/mega65/mouse.h
