+++ include/mega65/tests.h
