+++ src/mouse.c
@@ -19,7 +19,8 @@
     Date      2020-07-01
 */
 
-#include <memory.h>
+#include <mega65/memory.h>
+#include <mega65/mouse.h>
 
 unsigned short mouse_min_x = 0;
 unsigned short mouse_min_y = 0;
